utils dynamic query

+++ app/model/dev.go
@@ -1,7 +1,7 @@
 package model
 
 type Dev struct {
-	ID			int64 	`json:"id,omitempty" col:"id"`
+	ID			int 	`json:"id,omitempty" col:"id"`
 	Name 		string 	`json:"name,omitempty" col:"name"`
 	Email 		string 	`json:"email,omitempty" col:"email"`
 	Username	string 	`json:"username,omitempty" col:"username"`
